Rename downloadCredentials to hostCredentials
